Merge the latest code into master

+++ exchangis-job/exchangis-job-server/src/main/java/com/webank/wedatasphere/exchangis/job/server/restful/execute/ExchangisJobExecuteRestfulApi.java
@@ -1,13 +1,19 @@
 package com.webank.wedatasphere.exchangis.job.server.restful.execute;
 
+import com.webank.wedatasphere.exchangis.common.AuditLogUtils;
+import com.webank.wedatasphere.exchangis.common.UserUtils;
+import com.webank.wedatasphere.exchangis.common.enums.OperateTypeEnum;
+import com.webank.wedatasphere.exchangis.common.enums.TargetTypeEnum;
 import com.webank.wedatasphere.exchangis.datasource.core.utils.Json;
 import com.webank.wedatasphere.exchangis.job.domain.ExchangisJobInfo;
+import com.webank.wedatasphere.exchangis.job.domain.OperationType;
+import com.webank.wedatasphere.exchangis.job.launcher.ExchangisLauncherConfiguration;
 import com.webank.wedatasphere.exchangis.job.launcher.domain.task.TaskStatus;
 import com.webank.wedatasphere.exchangis.job.log.LogQuery;
 import com.webank.wedatasphere.exchangis.job.server.exception.ExchangisJobServerException;
 import com.webank.wedatasphere.exchangis.job.server.service.JobInfoService;
 import com.webank.wedatasphere.exchangis.job.server.service.impl.DefaultJobExecuteService;
-import com.webank.wedatasphere.exchangis.job.server.utils.AuthorityUtils;
+import com.webank.wedatasphere.exchangis.job.server.utils.JobAuthorityUtils;
 import com.webank.wedatasphere.exchangis.job.server.vo.ExchangisCategoryLogVo;
 import com.webank.wedatasphere.exchangis.job.server.vo.ExchangisJobProgressVo;
 import com.webank.wedatasphere.exchangis.job.server.vo.ExchangisJobTaskVo;
@@ -48,7 +54,8 @@ public class ExchangisJobExecuteRestfulApi {
     @RequestMapping( value = "/{id}/execute", method = RequestMethod.POST)
     public Message executeJob(@RequestBody(required = false) Map<String, Object> permitPartialFailures,
                               @PathVariable("id") Long id, HttpServletRequest request) {
-        String loginUser = SecurityFilter.getLoginUsername(request);
+        String loginUser = UserUtils.getLoginUser(request);
+        String oringinUser = SecurityFilter.getLoginUsername(request);
         Message result = Message.ok("Submitted succeed(提交成功)！");
         ExchangisJobInfo jobInfo = null;
         try {
@@ -60,12 +67,10 @@ public class ExchangisJobExecuteRestfulApi {
             // Convert to the job info
             jobInfo = new ExchangisJobInfo(jobVo);
 
-            if (!hasAuthority(loginUser, jobInfo)){
+            if (!JobAuthorityUtils.hasJobAuthority(loginUser, id, OperationType.JOB_EXECUTE)){
                 return Message.error("You have no permission to execute job (没有执行任务权限)");
             }
-            /*if (!AuthorityUtils.hasOwnAuthority(jobVo.getProjectId(), loginUser) && !AuthorityUtils.hasExecAuthority(jobVo.getProjectId(), loginUser)) {
-                return Message.error("You have no permission to execute job (没有执行任务权限)");
-            }*/
+
             // Send to execute service
             String jobExecutionId = executeService.executeJob(jobInfo, StringUtils.isNotBlank(jobInfo.getExecuteUser()) ?
                     jobInfo.getExecuteUser() : loginUser);
@@ -82,13 +87,19 @@ public class ExchangisJobExecuteRestfulApi {
             LOG.error(message, e);
         }
         result.setMethod("/api/rest_j/v1/dss/exchangis/main/job/{id}/execute");
+        assert jobInfo != null;
+        AuditLogUtils.printLog(oringinUser, loginUser, TargetTypeEnum.JOB, id.toString(), "Execute task is: " + jobInfo.getName(), OperateTypeEnum.EXECUTE, request);
         return result;
     }
 
     @RequestMapping( value = "/execution/{jobExecutionId}/taskList", method = RequestMethod.GET)
-    public Message getExecutedJobTaskList(@PathVariable(value = "jobExecutionId") String jobExecutionId) {
+    public Message getExecutedJobTaskList(@PathVariable(value = "jobExecutionId") String jobExecutionId, HttpServletRequest request) {
         Message message = Message.ok("Submitted succeed(提交成功)！");
+        String loginUser = UserUtils.getLoginUser(request);
         try {
+            if(!JobAuthorityUtils.hasJobExecuteSituationAuthority(loginUser, jobExecutionId, OperationType.JOB_QUERY)) {
+                return Message.error("You have no permission to get taskList (没有获取任务列表权限)");
+            }
             List<ExchangisJobTaskVo> jobTaskList = executeService.getExecutedJobTaskList(jobExecutionId);
             message.data("tasks", jobTaskList);
         } catch (ExchangisJobServerException e) {
@@ -96,14 +107,18 @@ public class ExchangisJobExecuteRestfulApi {
             LOG.error(errorMessage, e);
             message = Message.error(message + "(执行任务出错), reason: " + e.getMessage());
         }
-        message.setMethod("/api/rest_j/v1/dss/exchangis/main/job/execution/" + jobExecutionId + "/taskList");
+        message.setMethod("/api/rest_j/v1/" + jobExecutionId + "/taskList");
         return message;
     }
 
     @RequestMapping( value = "/execution/{jobExecutionId}/progress", method = RequestMethod.GET)
-    public Message getExecutedJobAndTaskStatus(@PathVariable(value = "jobExecutionId") String jobExecutionId) {
+    public Message getExecutedJobAndTaskStatus(@PathVariable(value = "jobExecutionId") String jobExecutionId, HttpServletRequest request) {
         ExchangisJobProgressVo jobAndTaskStatus;
+        String loginUser = UserUtils.getLoginUser(request);
         try {
+            if(!JobAuthorityUtils.hasJobExecuteSituationAuthority(loginUser, jobExecutionId, OperationType.JOB_QUERY)) {
+                return Message.error("You have no permission to get task progress (没有获取任务进度权限)");
+            }
             jobAndTaskStatus = executeService.getExecutedJobProgressInfo(jobExecutionId);
         } catch (ExchangisJobServerException e) {
             // TODO Log exception
@@ -116,9 +131,13 @@ public class ExchangisJobExecuteRestfulApi {
     }
 
     @RequestMapping( value = "/execution/{jobExecutionId}/status", method = RequestMethod.GET)
-    public Message getExecutedJobStatus(@PathVariable(value = "jobExecutionId") String jobExecutionId) {
+    public Message getExecutedJobStatus(@PathVariable(value = "jobExecutionId") String jobExecutionId, HttpServletRequest request) {
         Message message = Message.ok("Submitted succeed(提交成功)！");
+        String loginUser = UserUtils.getLoginUser(request);
         try {
+            if(!JobAuthorityUtils.hasJobExecuteSituationAuthority(loginUser, jobExecutionId, OperationType.JOB_QUERY)) {
+                return Message.error("You have no permission to get tastStatus (没有权限去获取任务状态)");
+            }
             ExchangisJobProgressVo jobStatus = executeService.getJobStatus(jobExecutionId);
             message.setMethod("/api/rest_j/v1/dss/exchangis/main/job/execution/" + jobExecutionId + "/status");
             message.data("status", jobStatus.getStatus());
@@ -143,9 +162,14 @@ public class ExchangisJobExecuteRestfulApi {
         Message result = Message.ok("Submitted succeed(提交成功)！");
         LogQuery logQuery = new LogQuery(fromLine, pageSize,
                 ignoreKeywords, onlyKeywords, lastRows);
+        String loginUser = UserUtils.getLoginUser(request);
         try {
+            if(!JobAuthorityUtils.hasJobExecuteSituationAuthority(loginUser, jobExecutionId, OperationType.JOB_QUERY)) {
+                return Message.error("You have no permission to get logs (没有获取任务日志权限)");
+            }
+
             ExchangisCategoryLogVo categoryLogVo = this.executeService
-                    .getJobLogInfo(jobExecutionId, logQuery, SecurityFilter.getLoginUsername(request));
+                    .getJobLogInfo(jobExecutionId, logQuery);
             result.setData(Json.convert(categoryLogVo, Map.class, String.class, Object.class));
         } catch (ExchangisJobServerException e) {
             String message = "Error occur while querying job log: [job_execution_id: " + jobExecutionId  +"]";
@@ -157,10 +181,14 @@ public class ExchangisJobExecuteRestfulApi {
     }
 
     @RequestMapping( value = "/execution/{jobExecutionId}/kill", method = RequestMethod.POST)
-    public Message ExecutedJobKill(@PathVariable(value = "jobExecutionId") String jobExecutionId) throws ExchangisJobServerException {
+    public Message ExecutedJobKill(@PathVariable(value = "jobExecutionId") String jobExecutionId, HttpServletRequest request) throws ExchangisJobServerException {
         ExchangisJobProgressVo jobStatus = executeService.getJobStatus(jobExecutionId);
-        String status = jobStatus.getStatus().toString();
         Message message = null;
+        String loginUser = SecurityFilter.getLoginUsername(request);
+        String oringinUser = SecurityFilter.getLoginUsername(request);
+        if(!JobAuthorityUtils.hasJobExecuteSituationAuthority(loginUser, jobExecutionId, OperationType.JOB_EXECUTE)) {
+            return Message.error("You have no permission to get kill job (没有权限去杀死任务)");
+        }
         if (!TaskStatus.isCompleted(jobStatus.getStatus()))
         {
             message = Message.ok("Kill succeed(停止成功)！");
@@ -176,6 +204,7 @@ public class ExchangisJobExecuteRestfulApi {
             message = Message.error("Kill failed(停止失败)！,job 已经到终态");
         }
         message.setMethod("/api/rest_j/v1/dss/exchangis/main/job/execution/" + jobExecutionId + "/kill");
+        AuditLogUtils.printLog(oringinUser, loginUser, TargetTypeEnum.JOB, jobExecutionId, "Kill job: ", OperateTypeEnum.KILL, request);
         return message;
     }
 
@@ -189,6 +218,7 @@ public class ExchangisJobExecuteRestfulApi {
                              @RequestParam(value = "size", required = false) int size,
                             HttpServletRequest request) {
         Message message = Message.ok("Submitted succeed(提交成功)！");
+        jobName = jobName.replace("_", "\\_");
         try {
             List<ExchangisLaunchedJobListVo> jobList = executeService.getExecutedJobList(jobExecutionId, jobName, status,
                     launchStartTime, launchEndTime, current, size, request);
@@ -204,21 +234,17 @@ public class ExchangisJobExecuteRestfulApi {
         return message;
     }
 
-    /**
-     * TODO complete the authority strategy
-     * @param username username
-     * @param jobInfo job info
-     * @return
-     */
-    private boolean hasAuthority(String username, ExchangisJobInfo jobInfo){
-        return username.equals(jobInfo.getCreateUser());
-    }
-
     @RequestMapping( value = "/{jobExecutionId}/deleteJob", method = RequestMethod.POST)
-    public Message ExecutedJobDelete(@PathVariable(value = "jobExecutionId") String jobExecutionId) throws ExchangisJobServerException {
-        //ExchangisLaunchedJobEntity jobAndTaskStatus = exchangisExecutionService.getExecutedJobAndTaskStatus(jobExecutionId);
+    public Message ExecutedJobDelete(@PathVariable(value = "jobExecutionId") String jobExecutionId, HttpServletRequest request) throws ExchangisJobServerException {
+        if (ExchangisLauncherConfiguration.LIMIT_INTERFACE.getValue()) {
+            return Message.error("You have no permission to delete this record (没有删除历史记录权限)");
+        }
         Message message = Message.ok("Kill succeed(停止成功)！");
+        String loginUser = UserUtils.getLoginUser(request);
         try {
+            if(!JobAuthorityUtils.hasJobExecuteSituationAuthority(loginUser, jobExecutionId, OperationType.JOB_EXECUTE)) {
+                return Message.error("You have no permission to delete this record (没有删除历史记录权限)");
+            }
             executeService.deleteJob(jobExecutionId);
             message.data("jobExecutionId", jobExecutionId);
         } catch (ExchangisJobServerException e){
@@ -231,10 +257,14 @@ public class ExchangisJobExecuteRestfulApi {
     }
 
     @RequestMapping( value = "/{jobExecutionId}/allTaskStatus", method = RequestMethod.GET)
-    public Message allTaskStatus(@PathVariable(value = "jobExecutionId") String jobExecutionId) throws ExchangisJobServerException {
+    public Message allTaskStatus(@PathVariable(value = "jobExecutionId") String jobExecutionId, HttpServletRequest request) throws ExchangisJobServerException {
         //ExchangisLaunchedJobEntity jobAndTaskStatus = exchangisExecutionService.getExecutedJobAndTaskStatus(jobExecutionId);
         Message message = Message.ok("所有任务状态");
+        String loginUser = UserUtils.getLoginUser(request);
         try {
+            if(!JobAuthorityUtils.hasJobExecuteSituationAuthority(loginUser, jobExecutionId, OperationType.JOB_QUERY)) {
+                return Message.error("You have no permission to get tastStatus (没有权限去获取任务状态)");
+            }
             List<String> allStatus = executeService.allTaskStatus(jobExecutionId);
             message.data("allStatus", allStatus);
             message.data("jobExecutionId", jobExecutionId);
